pyproject.toml
[tool.poetry]
name = "aikg"
version = "0.1.0"
description = "Langchain-powered natural language interface to RDF knowledge-graphs"
authors = ["cyril.matthey-doret <cyril.matthey-doret@epfl.ch>"]
license = "MIT"
readme = "README.md"
packages = [{include = "aikg"}]

[tool.poetry.dependencies]
python = "^3.10"
metaflow = "^2.8.3"
pyllamacpp = "^1.0.6"
tqdm = "^4.65.0"
llama-cpp-python = "^0.1.35"
transformers = "^4.28.1"
python-dotenv = "^1.0.0"
ipykernel = "^6.22.0"
llama-index = "0.6.0.alpha3"
rdflib = "6.2.0"
sentence-transformers = "^2.2.2"
chromadb = "^0.3.21"
more-itertools = "^9.1.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.3.1"
pre-commit = "^3.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
